// src/integrations/supabase/types.ts
export interface Database {
  public: {
    Tables: {
// ...
      payment_transactions: {
        Row: {
          id: string
          user_email: string
          amount: number
          method: string
          trx_id: string
          sender_number: string
          status: string | null
          created_at: string | null
        }
        Insert: {
          id?: string
          user_email: string
          amount: number
          method: string
          trx_id: string
          sender_number: string
          status?: string | null
          created_at?: string | null
        }
        Update: {
          id?: string
          user_email?: string
          amount?: number
          method?: string
          trx_id?: string
          sender_number?: string
          status?: string | null
          created_at?: string | null
        }
      }
      app_users: {
        Row: {
          id: number
          key: string
          pas: string | null
        }
        Insert: {
          id?: never
          key: string
          pas?: string | null
        }
        Update: {
          id?: never
          key?: string
          pas?: string | null
        }
      }
      referral_codes: {
        Row: {
          id: string
          code: string
          type: string
          value: number
          status: string
          created_at: string
        }
        Insert: {
          id?: string
          code: string
          type: string
          value: number
          status?: string
          created_at?: string
        }
        Update: {
          id?: string
          code?: string
          type?: string
          value?: number
          status?: string
          created_at?: string
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}